pyproject.toml
[project]
name = "argus"
version = "6.6.1"
description = "Argus - 热点新闻聚合与分析工具"
requires-python = ">=3.12"
dependencies = [
    "requests==2.33.0",
    "pytz==2026.1",
    "PyYAML==6.0.3",
    "fastmcp==2.12.5",
    "websockets==13.1",
    "feedparser==6.0.12",
    "boto3==1.42.76",
    "litellm==1.82.6",
    "json-repair==0.58.6",
    "tenacity==8.5.0",
    "jieba>=0.42.1",
    "rank-bm25>=0.2.2",
    "numpy>=2.4.4",
    "ruamel-yaml>=0.19.1",
]

[project.scripts]
argus = "argus.__main__:main"
argus-mcp = "argus_server.server:run_server"

[dependency-groups]
dev = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["argus", "argus_server"]
